refactor miot device adapter
